pyproject.toml
[tool.poetry]
name = "group-project-griffith"
version = "0.1.0"
description = "Griffith BSC 3 year group project"
authors = ["Thomas ODIN <thomasodin13@gmail.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
torch = ">=2.0.0"
transformers = ">=4.36.0"
sentence-transformers = ">=2.2.2"
tqdm = ">=4.66.1"
accelerate = ">=0.23.0"
scikit-learn = ">=1.3.0"
pymupdf = "^1.26.0"
langchain = "^0.3.25"
black = "^25.1.0"
vulture = "^2.14"
python-dotenv = "^1.1.0"
pinecone = "^7.0.1"
faiss-cpu = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
